implemented authorization and validation middleware

+++ src/controllers/expense.controller.js
@@ -1,18 +1,29 @@
 const {pool} = require('../database/db')
 
-const getAllExpense =async (req,res) =>{
+const getAllExpense = async (req,res) =>{
     try {
-        const result = await pool.query("SELECT * FROM expenses");
-        return res.status(200).json({"expenses":result.rows})
+        const user_id = req.user;
+        let total_expense = 0;
+        const resultQuery = await pool.query("SELECT * FROM expenses WHERE user_id=$1",[user_id]);
+        const results = resultQuery.rows;
+        total_expense = results.reduce((accumulator,result)=>{
+            return accumulator + parseFloat(result.amount);
+        },0)
+        return res.status(200).json({"expenses":results,"total_expenses":total_expense})
     } catch (error) {
-        return res.status(500).json({"message":"Error fetching expenses",error})
+        console.error("Error fetching expenses",error)
+        return res.status(500).json({"message":"Error fetching expenses"})
     }
 }
 
 const addExpense = (req,res) =>{
     try {
+        const user_id = req.user;
         const {category,description,amount,date} = req.body;
-        const result = pool.query("INSERT INTO expenses (category,description,amount,date) VALUES ($1,$2,$3,$4)",[category,description,amount,date]);
+        const resultQuery = pool.query("INSERT INTO expenses (category,description,amount,date,user_id) VALUES ($1,$2,$3,$4,$5)",[category,description,amount,date,user_id]);
+        if(resultQuery.rowCount===0){
+            return res.status(400).json({"message":"record expense failed !"})
+        }
         return res.status(200).json({"message":"Data added successfully"})
     } catch (error) {
         return res.status(500).json({"message":"Failed to record"})
@@ -23,10 +34,16 @@ const addExpense = (req,res) =>{
 
 const updateExpense = async (req,res) =>{
     try{
+        const user_id = req.user;
         const {id} = req.params;
         const idExist = await pool.query("SELECT * FROM expenses WHERE id=$1",[id])
         if(idExist.rows.length===0){
             return res.status(404).json({"message":"No expense by that id"})
+        }
+        const expense = idExist.rows[0];
+        const isAuthorised = expense.user_id === user_id;
+        if(!isAuthorised){
+            return res.status(403).json({ "message": "You are not authorized to update this expense" }); 
         }
         const {category,description,amount,date} = req.body;
         
@@ -59,10 +76,16 @@ const updateExpense = async (req,res) =>{
 
 const deleteExpense = async (req,res) =>{
     try {
+        const user_id = req.user;
         const {id} = req.params;
         const idExist = await pool.query("SELECT * FROM expenses WHERE id=$1",[id])
         if(idExist.rows.length===0){
             return res.status(404).json({"message":"No expense by that id"})
+        }
+        const expense = idExist.rows[0];
+        const isAuthorised = expense.user_id === user_id;
+        if(!isAuthorised){
+            return res.status(403).json({ "message": "You are not authorized to access this expense" }); 
         }
         const deleteQuery = await pool.query("DELETE FROM expenses WHERE id=$1",[id]);
         if(deleteQuery.rowCount===0){
